+++ CRM.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 15.0.27130.2036
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CRM", "CRM\CRM.csproj", "{21BF54AC-186A-4C26-8599-238BE6634E17}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CRMData", "CRMData\CRMData.csproj", "{B28269B4-88BC-4330-8336-CC63F5312E89}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CRM.DAL", "CRMData\CRM.DAL.csproj", "{B28269B4-88BC-4330-8336-CC63F5312E89}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
