+++ CRM/CRM.csproj
@@ -308,9 +308,9 @@
     <Content Include="Scripts\jquery-1.10.2.min.map" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\CRMData\CRMData.csproj">
+    <ProjectReference Include="..\CRMData\CRM.DAL.csproj">
       <Project>{B28269B4-88BC-4330-8336-CC63F5312E89}</Project>
-      <Name>CRMData</Name>
+      <Name>CRM.DAL</Name>
     </ProjectReference>
   </ItemGroup>
   <PropertyGroup>
