+++ CRMData/Adapters/Extentions/QueryCustomerExtentions.cs
@@ -1,7 +1,7 @@
 ﻿using CRMData.Entities;
 using System.Linq;
 
-namespace CRMData.Adapters.Extentions
+namespace CRM.DAL.Adapters.Extentions
 {
     public static class QueryUserExtentions
     {
