+++ CRMData/Adapters/LeadAdapter.cs
@@ -1,9 +1,9 @@
-﻿using CRMData.Contexts;
-using CRMData.Entities;
+﻿using CRM.DAL.Contexts;
+using CRM.DAL.Entities;
 using System.Collections.Generic;
 using System.Linq;
 using System.Data.Entity;
-using CRMData.Adapters.Extentions;
+using CRM.DAL.Adapters.Extentions;
 
 namespace CRMData.Adapters
 {
