+++ CRMData/CRM.DAL.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{B28269B4-88BC-4330-8336-CC63F5312E89}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>CRMData</RootNamespace>
-    <AssemblyName>CRMData</AssemblyName>
+    <RootNamespace>CRM.DAL</RootNamespace>
+    <AssemblyName>CRM.DAL</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -86,6 +86,8 @@
     <Compile Include="Adapters\Extentions\QueryLeadExtentions.cs" />
     <Compile Include="Adapters\LeadAdapter.cs" />
     <Compile Include="Contexts\BaseContext.cs" />
+    <Compile Include="Contexts\ContextFactory.cs" />
+    <Compile Include="Contexts\CRMDBInitializer.cs" />
     <Compile Include="Entities\Address.cs" />
     <Compile Include="Entities\Call.cs" />
     <Compile Include="Entities\DAddressType.cs" />
